verification add name of person who verified in direct message

+++ cogs/verification.py
@@ -25,25 +25,24 @@ class Verification(commands.Cog, name="Verification"):
             return
         if payload is not None:
             search = await WorldRecords.find_one({"message_id": payload.message_id})
-            if search is not None:
-                if payload.message_id == search.message_id:
-                    guild = self.bot.get_guild(payload.guild_id)
-                    channel = guild.get_channel(payload.channel_id)
-                    msg = await channel.fetch_message(payload.message_id)
-                    hidden_channel = guild.get_channel(constants.HIDDEN_VERIFICATION_CHANNEL)
-                    hidden_msg = await hidden_channel.fetch_message(search.hidden_id)
-                    await hidden_msg.delete()
-                    if str(payload.emoji) == constants.VERIFIED_EMOJI:
-                        search.verified = True
-                        await search.commit()
-                        await msg.author.send(f'Your submission has been verified!\n```Map Code: {search.code}{constants.NEW_LINE}Level: {search.level}{constants.NEW_LINE}Record: {utilities.display_record(search.record)}```{msg.jump_url}')
+            if search is not None and payload.message_id == search.message_id:
+                guild = self.bot.get_guild(payload.guild_id)
+                channel = guild.get_channel(payload.channel_id)
+                msg = await channel.fetch_message(payload.message_id)
+                hidden_channel = guild.get_channel(constants.HIDDEN_VERIFICATION_CHANNEL)
+                hidden_msg = await hidden_channel.fetch_message(search.hidden_id)
+                await hidden_msg.delete()
+                if str(payload.emoji) == constants.VERIFIED_EMOJI:
+                    search.verified = True
+                    await search.commit()
+                    await msg.author.send(f'Your submission has been verified by {payload.member.name}!\n```Map Code: {search.code}{constants.NEW_LINE}Level: {search.level}{constants.NEW_LINE}Record: {utilities.display_record(search.record)}```{msg.jump_url}')
 
-                    elif str(payload.emoji) == constants.NOT_VERIFIED_EMOJI:
-                        search.verified = False
-                        await search.commit()
-                        await msg.author.send(f'Your submission failed verification!\n```Map Code: {search.code}{constants.NEW_LINE}Level: {search.level}{constants.NEW_LINE}Record: {utilities.display_record(search.record)}```{msg.jump_url}')
+                elif str(payload.emoji) == constants.NOT_VERIFIED_EMOJI:
+                    search.verified = False
+                    await search.commit()
+                    await msg.author.send(f'{payload.member.name} has denied verification for your submission! {payload.member.name}!\n```Map Code: {search.code}{constants.NEW_LINE}Level: {search.level}{constants.NEW_LINE}Record: {utilities.display_record(search.record)}```{msg.jump_url}')
 
-                    await msg.clear_reactions()
+                await msg.clear_reactions()
 
     @commands.Cog.listener()
     async def on_raw_message_delete(self, payload=None):
